+++ packages/fighting-design/_hooks/use-list/index.ts
@@ -71,6 +71,35 @@ export const useList = <T extends object>(prop: T, name: string): UseListReturn 
   }
 
   /**
+ * 设置样式列表的值
+ *
+ * @param { string | number } val 值
+ * @param { string } key 键
+ * @returns { string | number } 处理后的值
+ */
+  const setListValue = (val: string | number, key: string, pixel: boolean | string | string[] = true): string | number => {
+    /** 如果需要添加单位，则所有的数字都添加单位 */
+    if (isBoolean(pixel)) {
+      return (isNumber(val) ? (pixel ? sizeChange(val) : val) : val) as
+        | string
+        | number
+    } else if (isString(pixel)) {
+      /** 如果为字符串类型，则代表仅仅有一个不需要添加单位 */
+      if (pixel === key) return val
+    } else if (isArray(pixel)) {
+      /** 如果为数组类型，则代表有些值不需要添加单位，循环遍历处理 */
+      pixel.forEach((item: string): void | unknown => {
+        if (item === val) {
+          return val
+        }
+      })
+    }
+
+    /** 没有进入判断则默认添加 */
+    return sizeChange(val)
+  }
+
+  /**
    * 样式列表
    *
    * @param { Array } list 样式所需要的 prop 参数
@@ -87,51 +116,12 @@ export const useList = <T extends object>(prop: T, name: string): UseListReturn 
       /** 过滤得到 prop 集合 */
       const propList: Record<string, unknown> = filter(list)
 
-      /**
-       * 设置样式列表的值
-       *
-       * @param { string | number } val 值
-       * @param { string } key 键
-       * @returns { string | number } 处理后的值
-       */
-      const setListValue = (val: string | number, key: string): string | number => {
-        /** 如果需要添加单位，则所有的数字都添加单位 */
-        if (isBoolean(pixel)) {
-          return (isNumber(val) ? (pixel ? sizeChange(val) : val) : val) as
-            | string
-            | number
-        } else if (isString(pixel)) {
-          /** 如果为字符串类型，则代表仅仅有一个不需要添加单位 */
-          if (pixel === key) return val
-        } else if (isArray(pixel)) {
-          /** 如果为数组类型，则代表有些值不需要添加单位，循环遍历处理 */
-          pixel.forEach((item: string): void | unknown => {
-            if (item === val) {
-              return val
-            }
-          })
-        }
-
-        /** 没有进入判断则默认添加 */
-        return sizeChange(val)
-      }
-
       for (const key in propList) {
         if (propList[key]) {
-          /**
-           * @description 为什么要进行 isNumber 判断？
-           *
-           * 因为很多属性是同时支持 number 和 staring 类型的参数
-           *
-           * 所以这里要进行判断，如果是数字类型，则需要使用 sizeChange 方法进行转换标注单位
-           *
-           * @description convertFormat 方法描述
-           *
-           * 因为 prop 参数的键都是驼峰命名法，所以这里要转换为短横线连接命名
-           */
           styleList[`--${name}-${convertFormat(key)}`] = setListValue(
             propList[key] as string | number,
-            key
+            key,
+            pixel
           )
         }
       }