+++ packages/fighting-design/image/src/image.vue
@@ -1,15 +1,13 @@
 <script lang="ts" setup name="FImage">
   import { Props } from './props'
-  import { onMounted, ref, computed } from 'vue'
-  import { sizeChange } from '../../_utils'
-  import { useLoadImg, useProps } from '../../_hooks'
-  import type { CSSProperties } from 'vue'
-  import type { ClassList } from '../../_interface'
+  import { onMounted, ref } from 'vue'
+  import { useLoadImg, useProps, useList } from '../../_hooks'
   import type { UseLoadImgProp } from '../../_hooks'
 
   const prop = defineProps(Props)
 
   const { filter } = useProps(prop)
+  const { classes, styles } = useList(prop, 'image')
   const { loadImg, isSuccess, isShowNode } = useLoadImg(
     filter([
       'src',
@@ -29,43 +27,20 @@
   })
 
   /** 类名列表 */
-  const classList = computed((): ClassList => {
-    const { fit, noSelect } = prop
-
-    return [
-      'f-image__img',
-      {
-        [`f-image__${fit}`]: fit,
-        'f-image__select': noSelect
-      }
-    ]
-  })
+  const classList = classes(['fit', 'select', 'block'], 'f-image')
 
   /** 样式列表 */
-  const styleList = computed((): CSSProperties => {
-    const { width, height, round } = prop
-
-    return {
-      '--image-width': sizeChange(width),
-      '--image-height': sizeChange(height),
-      '--image-border-radius': sizeChange(round)
-    } as CSSProperties
-  })
+  const styleList = styles(['width', 'height', 'round'])
 </script>
 
 <template>
-  <div
-    v-if="isSuccess"
-    role="img"
-    :class="['f-image', { 'f-image__block': block }]"
-    :style="styleList"
-  >
+  <div v-if="isSuccess" role="img" :class="classList" :style="styleList">
     <!-- 真正展示的图片 -->
     <img
       ref="imageEl"
+      class="f-image__img"
       src=""
-      :class="classList"
-      :style="(styleList, isShowNode ? '' : 'visibility: hidden')"
+      :style="isShowNode ? '' : 'visibility: hidden'"
       :draggable="draggable"
       :referrer-policy="referrerPolicy"
       :alt="alt"