+++ start/src/App.vue
@@ -1,5 +1,8 @@
 <script lang="ts" setup></script>
 
-<template></template>
+<template>
+  <f-image width="200px" :round="13" block src="https://tianyuhao.cn/images/auto/1.jpg" />
+  123
+</template>
 
 <style lang="scss" scoped></style>