+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- `f-image` 组件 `no-select` 配置项改为 `select`
+
 ## 0.27.2 (2023-03-13)
 
 **修复**